+++ Cargo.toml
@@ -2,8 +2,14 @@
 members = [
     ".",
     "lib/hyper-balance",
+    "lib/linkerd2-addr",
+    "lib/linkerd2-conditional",
+    "lib/linkerd2-dns-name",
+    "lib/linkerd2-drain",
+    "lib/linkerd2-identity",
     "lib/linkerd2-metrics",
     "lib/linkerd2-router",
+    "lib/linkerd2-signal",
     "lib/linkerd2-stack",
     "lib/linkerd2-task",
     "lib/linkerd2-timeout",
@@ -22,13 +28,19 @@ default = ["flaky_tests"]
 flaky_tests = []
 
 [dependencies]
-hyper-balance      = { path = "lib/hyper-balance" }
-linkerd2-metrics   = { path = "lib/linkerd2-metrics" }
-linkerd2-never     = { path = "lib/linkerd2-never" }
-linkerd2-router    = { path = "lib/linkerd2-router" }
-linkerd2-stack     = { path = "lib/linkerd2-stack" }
-linkerd2-task      = { path = "lib/linkerd2-task" }
-linkerd2-timeout   = { path = "lib/linkerd2-timeout" }
+hyper-balance = { path = "lib/hyper-balance" }
+linkerd2-addr        = { path = "lib/linkerd2-addr" }
+linkerd2-conditional = { path = "lib/linkerd2-conditional" }
+linkerd2-dns-name    = { path = "lib/linkerd2-dns-name" }
+linkerd2-drain       = { path = "lib/linkerd2-drain" }
+linkerd2-identity    = { path = "lib/linkerd2-identity" }
+linkerd2-metrics     = { path = "lib/linkerd2-metrics" }
+linkerd2-never       = { path = "lib/linkerd2-never" }
+linkerd2-router      = { path = "lib/linkerd2-router" }
+linkerd2-signal      = { path = "lib/linkerd2-signal" }
+linkerd2-stack       = { path = "lib/linkerd2-stack" }
+linkerd2-task        = { path = "lib/linkerd2-task" }
+linkerd2-timeout     = { path = "lib/linkerd2-timeout" }
 
 linkerd2-proxy-api = { git = "https://github.com/linkerd/linkerd2-proxy-api", rev = "ddbc3a4f7f8b0058801f896d27974d19ee98094c" }
 
@@ -52,7 +64,6 @@ regex = "1.0.0"
 
 # networking
 tokio = "0.1.14"
-tokio-signal = "0.2"
 tokio-timer = "0.2.6"   # for tokio_timer::clock
 tower = "0.1"
 tower-discover = "0.1"
