+++ lib/linkerd2-addr/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "linkerd2-addr"
+version = "0.1.0"
+authors = ["Oliver Gould <ver@buoyant.io>"]
+edition = "2018"
+publish = false
+
+[dependencies]
+http = "0.1"
+linkerd2-dns-name = { path = "../linkerd2-dns-name" }
