+++ lib/linkerd2-conditional/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "linkerd2-conditional"
+version = "0.1.0"
+authors = ["Oliver Gould <ver@buoyant.io>"]
+edition = "2018"
+
+[dependencies]
