+++ lib/linkerd2-drain/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "linkerd2-drain"
+version = "0.1.0"
+authors = ["Oliver Gould <ver@buoyant.io>"]
+edition = "2018"
+
+[dependencies]
+futures = "0.1"
+linkerd2-never = { path = "../linkerd2-never" }
