+++ lib/linkerd2-identity/src/testdata/bar-ns1-ca1/crt.der
