+++ lib/linkerd2-identity/src/testdata/ca1-key.pem
