+++ lib/linkerd2-identity/src/testdata/controller-linkerd-ca1/crt.der
