+++ src/app/config.rs
@@ -2,8 +2,9 @@ use super::control::ControlAddr;
 use super::identity;
 use crate::proxy::reconnect::Backoff;
 use crate::transport::tls;
-use crate::{addr, dns, Addr, Conditional};
+use crate::{dns, Conditional};
 use indexmap::IndexSet;
+use linkerd2_addr::{self as addr, Addr};
 use std::collections::HashMap;
 use std::convert::TryFrom;
 use std::iter::FromIterator;
