+++ src/lib.rs
@@ -1,13 +1,9 @@
 #![deny(warnings, rust_2018_idioms)]
 #![recursion_limit = "128"]
 
-mod addr;
 pub mod app;
-mod conditional;
 pub mod control;
 mod dns;
-mod drain;
-mod identity;
 pub mod logging;
 mod proxy;
 mod svc;
@@ -15,8 +11,9 @@ mod tap;
 pub mod telemetry;
 pub mod transport;
 
-use self::addr::{Addr, NameAddr};
-use self::conditional::Conditional;
+use linkerd2_addr::{Addr, NameAddr};
+use linkerd2_conditional::Conditional;
+use linkerd2_identity as identity;
 
 pub use self::logging::trace;
 pub use self::transport::SoOriginalDst;
