+++ src/tap/grpc/server.rs
@@ -1,6 +1,7 @@
 use super::match_::Match;
 use crate::proxy::http::HasH2Reason;
 use crate::tap::{iface, Inspect};
+use crate::transport::tls;
 use crate::Conditional;
 use bytes::Buf;
 use futures::sync::mpsc;
@@ -427,8 +428,8 @@ fn base_event<B, I: Inspect>(req: &http::Request<B>, inspect: &I) -> api::TapEve
         source_meta: {
             let mut m = api::tap_event::EndpointMeta::default();
             let tls = inspect.src_tls(req);
-            let tls_status = tls.as_ref().map(|_| ()).to_string();
-            m.labels.insert("tls".to_owned(), tls_status);
+            let tls_status = tls::Status::from(tls.as_ref());
+            m.labels.insert("tls".to_owned(), tls_status.to_string());
             if let Conditional::Some(id) = tls {
                 m.labels
                     .insert("client_id".to_owned(), id.as_ref().to_owned());
@@ -441,8 +442,8 @@ fn base_event<B, I: Inspect>(req: &http::Request<B>, inspect: &I) -> api::TapEve
             m.labels
                 .extend(labels.iter().map(|(k, v)| (k.clone(), v.clone())));
             let tls = inspect.dst_tls(req);
-            let tls_status = tls.as_ref().map(|_| ()).to_string();
-            m.labels.insert("tls".to_owned(), tls_status);
+            let tls_status = tls::Status::from(tls.as_ref());
+            m.labels.insert("tls".to_owned(), tls_status.to_string());
             if let Conditional::Some(id) = tls {
                 m.labels
                     .insert("server_id".to_owned(), id.as_ref().to_owned());
